@@ -8,35 +8,41 @@
     "url": "http://fiahfy.blogspot.com/"
   },
   "dependencies": {
-    "@fiahfy/electron-accelerator-formatter": "^0.0.2",
-    "@fiahfy/electron-context-menu": "^0.0.2",
-    "@fiahfy/worker-promisify": "^0.0.3",
-    "@nuxtjs/vuetify": "^0.5.1",
-    "electron-devtools-installer": "^2.2.0",
-    "electron-window-state": "^5.0.0",
-    "file-url": "^2.0.2",
-    "material-design-icons-iconfont": "^4.0.2",
-    "nuxt": "^2.3.4",
-    "typeface-roboto": "^0.0.54",
+    "@fiahfy/electron-accelerator-formatter": "^0.0.4",
+    "@fiahfy/electron-context-menu": "^0.0.5",
+    "@fiahfy/worker-promisify": "^0.0.4",
+    "@mdi/font": "^5.1.45",
+    "@nuxtjs/vuetify": "^1.11.2",
+    "electron-devtools-installer": "^3.0.0",
+    "electron-window-state": "^5.0.3",
+    "file-url": "^3.0.0",
+    "material-design-icons-iconfont": "^5.0.1",
+    "nuxt": "^2.12.2",
+    "typeface-roboto": "^0.0.75",
     "vue-long-press-directive": "^1.0.2",
-    "vue-moment": "^4.0.0",
-    "vuex-persistedstate": "^2.0.0"
+    "vue-moment": "^4.1.0",
+    "vue-property-decorator": "^8.4.2",
+    "vuex-module-decorators": "^0.17.0",
+    "vuex-persistedstate": "^3.0.1"
   },
   "devDependencies": {
-    "@fiahfy/icns-convert": "^0.0.5",
-    "@fiahfy/ico-convert": "^0.0.4",
-    "cross-env": "^5.0.5",
-    "electron": "^4.0.0",
-    "electron-builder": "^20.0.4",
-    "eslint": "^5.5.0",
-    "eslint-config-prettier": "^4.0.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "eslint-plugin-vue": "^5.0.0-beta.3",
-    "node-sass": "^4.9.0",
-    "npm-run-all": "^4.0.2",
-    "prettier": "^1.14.3",
-    "sass-loader": "^7.0.1",
-    "vuetify-loader": "^1.1.1",
+    "@fiahfy/icns-convert": "^0.0.8",
+    "@fiahfy/ico-convert": "^0.0.7",
+    "@nuxt/typescript-build": "^0.6.6",
+    "@nuxtjs/eslint-config-typescript": "^1.0.2",
+    "@types/node": "12.12.25",
+    "cross-env": "^7.0.2",
+    "electron": "^8.2.5",
+    "electron-builder": "^22.6.0",
+    "eslint": "^7.0.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "husky": "^4.2.5",
+    "node-sass": "^4.14.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.5",
+    "sass-loader": "^8.0.2",
+    "vuetify-loader": "^1.4.3",
     "worker-loader": "^2.0.0"
   },
   "keywords": [
@@ -56,9 +62,8 @@
     "build": "nuxt generate",
     "dev": "cross-env NODE_ENV=development run-p serve electron",
     "electron": "electron main.js",
-    "lint": "eslint -f codeframe --ext .js,.vue --ignore-path .gitignore .",
-    "package": "electron-builder",
-    "release": "run-s build \"package -mw -p onTagOrDraft\"",
+    "lint": "eslint -f codeframe --ext .ts,.js,.vue --ignore-path .gitignore .",
+    "release": "electron-builder",
     "resources": "run-p resources:*",
     "resources:icns": "icns-convert build/icon.png",
     "resources:ico": "ico-convert build/icon.png",
