@@ -1,4 +1,4 @@
-const pkg = require('./package')
+import pkg from './package'
 
 module.exports = {
   /*
@@ -12,28 +12,29 @@ module.exports = {
         config.output.globalObject = 'this'
         config.module.rules.unshift({
           test: /\.worker\.js$/,
-          loader: 'worker-loader'
+          loader: 'worker-loader',
         })
       }
       // Set relative path
       config.output.publicPath = './_nuxt/'
-    }
+    },
   },
+
+  /*
+   ** Build modules
+   */
+  buildModules: ['@nuxt/typescript-build'],
 
   /*
    ** Global CSS
    */
-  css: [
-    'material-design-icons-iconfont/dist/material-design-icons.css',
-    'typeface-roboto/index.css',
-    '~/assets/css/app.css'
-  ],
+  css: ['@mdi/font/css/materialdesignicons.css', 'typeface-roboto/index.css'],
 
   /*
    ** Generate configuration
    */
   generate: {
-    dir: 'app'
+    dir: 'app',
   },
 
   /*
@@ -44,7 +45,10 @@ module.exports = {
   /*
    ** Headers of the page
    */
-  head: { title: pkg.productName },
+  head: {
+    title: pkg.productName,
+    meta: [{ hid: 'charset', charset: 'utf-8' }],
+  },
 
   /*
    ** SPA or Universal
@@ -58,13 +62,24 @@ module.exports = {
     [
       '@nuxtjs/vuetify',
       {
-        materialIcons: false,
+        customVariables: ['~/assets/variables.scss'],
+        defaultAssets: false,
         theme: {
-          primary: '#ff4081',
-          accent: '#ff4081'
-        }
-      }
-    ]
+          themes: {
+            light: {
+              primary: '#ff4081',
+              secondary: '#424242',
+              accent: '#ff4081',
+            },
+            dark: {
+              primary: '#ff4081',
+              secondary: '#E0E0E0',
+              accent: '#ff4081',
+            },
+          },
+        },
+      },
+    ],
   ],
 
   /*
@@ -76,14 +91,14 @@ module.exports = {
     '~/plugins/ipc-listener',
     '~/plugins/vue-long-press',
     '~/plugins/vue-moment',
-    '~/plugins/vuetify'
+    '~/plugins/vuetify',
   ],
 
   /*
    ** Router configuration
    */
   router: {
-    mode: 'hash'
+    mode: 'hash',
   },
 
   /*
@@ -96,7 +111,7 @@ module.exports = {
    */
   vue: {
     config: {
-      productionTip: false
-    }
-  }
+      productionTip: false,
+    },
+  },
 }
