@@ -1,5 +1,4 @@
 module.exports = {
   semi: false,
-  singleQuote: true,
-  arrowParens: 'always'
+  singleQuote: true
 }
