@@ -61,15 +61,15 @@ export default {
   props: {
     file: {
       type: Object,
-      default: () => ({})
-    }
+      default: () => ({}),
+    },
   },
   data() {
     return {
       loading: false,
       error: false,
       imageUrl: '',
-      images: ''
+      images: '',
     }
   },
   computed: {
@@ -80,18 +80,18 @@ export default {
       set(value) {
         this.$store.dispatch('local/explorer/updateFileRating', {
           filepath: this.file.path,
-          rating: value
+          rating: value,
         })
-      }
+      },
     },
     active() {
       return this.isFileSelected({ filepath: this.file.path })
     },
     icon() {
       if (this.file.exists) {
-        return this.file.directory ? 'folder' : 'photo'
+        return this.file.directory ? 'mdi-folder' : 'mdi-photo'
       }
-      return 'broken_image'
+      return 'mdi-broken_image'
     },
     iconColor() {
       if (this.file.exists) {
@@ -113,7 +113,7 @@ export default {
     },
     ...mapState('settings', ['thumbnailStyle']),
     ...mapGetters('settings', ['thumbnailHeightValue', 'isFileAvailable']),
-    ...mapGetters('local/explorer', ['isFileSelected'])
+    ...mapGetters('local/explorer', ['isFileSelected']),
   },
   async created() {
     if (!this.file.directory) {
@@ -123,7 +123,7 @@ export default {
     this.loading = true
     const { data } = await worker.postMessage({
       key: this.file.path,
-      data: this.file.path
+      data: this.file.path,
     })
     const filepathes = data.filter((filepath) =>
       this.isFileAvailable({ filepath })
@@ -147,8 +147,8 @@ export default {
         {
           label: 'View',
           click: () => this.viewFile({ filepath: this.file.path }),
-          accelerator: 'Enter'
-        }
+          accelerator: 'Enter',
+        },
       ]
       const text = getSelection().toString()
       if (text) {
@@ -159,11 +159,11 @@ export default {
           {
             label: `Search "${text}"`,
             click: () => this.searchFiles({ query: text }),
-            accelerator: 'CmdOrCtrl+F'
-          }
+            accelerator: 'CmdOrCtrl+F',
+          },
         ]
       }
-      this.$contextMenu.show(template)
+      this.$contextMenu.open(template)
     },
     onError() {
       this.error = true
@@ -172,9 +172,9 @@ export default {
       'selectFile',
       'searchFiles',
       'openFile',
-      'viewFile'
-    ])
-  }
+      'viewFile',
+    ]),
+  },
 }
 </script>
 
