@@ -2,5 +2,5 @@ export default {
   DIRECTORY_INPUT: 'input[name=directory]',
   QUERY_INPUT: 'input[name=query]',
   EXPLORER_TABLE: '.explorer-table',
-  EXPLORER_GRID_LIST: '.explorer-grid-list'
+  EXPLORER_GRID_LIST: '.explorer-grid-list',
 }
