@@ -8,19 +8,19 @@ const DEFAULT_EXTENSIONS = [
   'SVG',
   'TIF',
   'TIFF',
-  'WEBP'
+  'WEBP',
 ]
 
 const PREVIEW_WIDTHS = {
   none: 0,
   narrow: 128,
   medium: 192,
-  wide: 256
+  wide: 256,
 }
 const THUMBNAIL_HEIGHTS = {
   short: 128,
   medium: 192,
-  tall: 256
+  tall: 256,
 }
 
 const THUMBNAIL_STYLES = ['cover', 'contain']
@@ -29,5 +29,5 @@ export default {
   DEFAULT_EXTENSIONS,
   PREVIEW_WIDTHS,
   THUMBNAIL_HEIGHTS,
-  THUMBNAIL_STYLES
+  THUMBNAIL_STYLES,
 }
