@@ -1,4 +1,4 @@
 import Vue from 'vue'
-import show from '@fiahfy/electron-context-menu'
+import { open } from '@fiahfy/electron-context-menu'
 
-Vue.prototype.$contextMenu = { show }
+Vue.prototype.$contextMenu = { open }
